requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-disqus==0.5
django-htmlmin==0.10.0
django-nocaptcha-recaptcha==0.0.20
django-updown==1.1.0
djangorestframework==3.7.7
djipsum==1.1.5
Faker==0.8.11
html5lib==1.0.1
idna==2.6
Markdown==2.6.11
mysqlclient==1.4.2.post1
martor==1.3.1
mock==2.0.0
oauthlib==2.0.6
pbr==3.1.1
Pillow==5.0.0
pygeoip==0.3.2
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.22
webencodings==0.5.1


# for dummy
# djipsum==1.1.5
# Faker==0.8.11
# python-dateutil==2.6.1
# six==1.11.0
# text-unidecode==1.1
